+++ core/core-domain/src/main/java/com/imstargg/core/domain/statistics/BattleEventBrawlersResultStatistics.java
@@ -4,8 +4,8 @@ import java.util.List;
 
 public record BattleEventBrawlersResultStatistics(
         List<Long> brawlerBrawlStarsIds,
-        int victoryCount,
-        int defeatCount,
-        int drawCount
+        long victoryCount,
+        long defeatCount,
+        long drawCount
 ) {
 }
